// packages/rest/generated-defs/decorators.ts
export type ValidateHasKeyDecorator = (
  context: DecoratorContext,
  target: Type,
  value: Type
) => void;

export type ValidateIsErrorDecorator = (
  context: DecoratorContext,
  target: Type,
  value: Type
) => void;